requirements/base.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
django-cors-headers==3.12.0
django-environ==0.8.1